+++ app/src/main/java/com/example/tetrisg8/FichaView.java
@@ -19,7 +19,7 @@ public class FichaView extends View{
 
         public Pieza generarPieza (int x, int y){// la pieza necesita unas coordenadas por defecto en las que aparecer
             int tipoPieza = (int) Math.floor(Math.random()*6+1); //función para generar un numero aleatorio del 1 al 7
-            Pieza pieza = new Pieza (tipoPieza, x, y);
+            Pieza pieza = new Pieza (x, y, tipoPieza, 1);
             return pieza;
         }
